Enable composed characters
Fontin does not have all accented characters (e.g, ü or ç). When `compose=yes`
is added to default font features, ConTeXt uses fallback _composed_ characters
for these missing characters.

A discussion on composed characters is out of the scope of this example. Just
think of this as a workaround for an incomplete font.