Start a new page for each slide
Adding `page=yes` ensures that each slide starts a new page. If a slide
contains more material, it will overflow one page (and no title will appear on
the next page). So, it is the responsibility of the user to ensure that a slide
has the right amount of text.