Setup itemize head style
The itemize environment in ConTeXt provides two types of items: normal items
that are created using `\item` or `\startitem ... \stopitem` and items with
heads, created using `\head` or `\starthead ... \stophead`.

I like to use the second alternative in presentations. The advantage is that the
item head can be typeset in a different style than the regular items; so it adds
some emphasis to the main points.

See:

- http://wiki.contextgarden.net/Command/head

for more details.