Example text for images
The simpleslides module provides two ways of including images: a horizontal
mode and a vertical mode. I like that, and want to provide similar alternatives
in this case.

A horizontal image should occupy as must horizontal space as possible without
overflowing the page height.

A vertical image should not be wider than 0.5\textwidth and should occupy as
much vertical space as possible.

For now, we use dummy settings `horizontal` and `vertical` for these two types
of images. The actual settings will be defined in the the next commits.